move speed assigner to private header and singleton
