+++ src/mjolnir/CMakeLists.txt
@@ -30,13 +30,13 @@ if (UNIX)
     luatagtransform.cc
     osmdata.cc
     osmrestriction.cc
+    speed_assigner.h
     PROPERTIES COMPILE_FLAGS "-Werror -Wall")
 endif (UNIX)
 
 set(sources
   ${CMAKE_CURRENT_BINARY_DIR}/graph_lua_proc.h
   ${CMAKE_CURRENT_BINARY_DIR}/admin_lua_proc.h
-
   admin.cc
   adminbuilder.cc
   bssbuilder.cc
@@ -66,6 +66,7 @@ set(sources
   restrictionbuilder.cc
   servicedays.cc
   shortcutbuilder.cc
+  speed_assigner.h
   timeparsing.cc
   transitbuilder.cc
   util.cc
